requirements.txt
# Original packages from FinAgent
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
cbor==1.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
cryptography==46.0.3
curl_cffi==0.14.0
dataclasses-json==0.6.7
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
executing==2.2.1
filelock==3.20.1
finance-datareader==0.9.96
FlagEmbedding==1.3.5
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ijson==3.4.0.post0
inscriptis==2.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.58
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langgraph==0.3.2
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.74
langsmith==0.3.45
lxml==5.4.0
lz4==4.4.5
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pandas==2.3.3
parso==0.8.5
peewee==3.18.3
peft==0.18.0
pexpect==4.9.0
platformdirs==4.5.1
plotly==6.5.0
postgrest==2.24.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.13.1
tavily-python==0.7.13
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trec-car-tools==2.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unlzw3==0.2.3
urllib3==2.6.2
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zlib-state==0.1.10
zstandard==0.23.0

# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20